requirements.txt
asgiref==3.8.1
attrs==23.2.0
cattrs==23.2.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-schema-graph==3.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
Markdown==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
